requirements.txt
Django>=3.0.8,<3.2.0

djangorestframework>=3.11.0,<3.12.0

psycopg2>=2.7.5,<2.8.0

flake8>=3.8.3,<3.9.0

django-filter>=2.3.0,<2.5.0
